requirements.txt
fastapi[standard]==0.116.
pydantic==2.11.7
python-multipart==0.0.20
google-genai==1.31.0